integrate Room db

+++ app/src/main/java/com/example/offline/model/ModelConstants.java
@@ -1,5 +1,5 @@
 package com.example.offline.model;
 
 public class ModelConstants {
-    public static int DUMMY_PHOTO_ID = 1;
+    public static long DUMMY_PHOTO_ID = 1L;
 }
